Allow crit_section nesting in prestarted timeout
Due to a bug in critical section implementation, exiting critical
section can be preempted, causing rx_prestarted_timeout, which is called
from rtc_irq, to not be able to enter critical section normally if it
is the IRQ that preempts exiting critical section. Allowing for nesting
should temporarily solve the problem until the bug is fixed.