KRKNWK-7859: Clear pending RADIO IRQ in NVIC when disabling TRX
It might happen that while MARGIN event is being handled from
an interrupt with priority equal 0, RADIO event occurs before
RADIO peripheral is reset. It causes pending interrupt handled
after timeslot is ended by the 802.15.4 driver. This patch clears
any pending RADIO interrupts when resetting the peripheral, so that
the interrupt is ignored and the timeslot end is respected.